+++ vue/src/components/Dialogs/Setup/DeploymentSpecifications/UpdateDialogs/DeploymentSpecificationUpdateEnvironmentVariableDialog.vue
@@ -2,6 +2,7 @@
 import {computed, defineComponent, onMounted, onUnmounted, reactive, ref, watch} from 'vue'
 import type {DialogEventsInterface} from "@/components/Dialogs/DialogEventsInterface";
 import {VTextField} from "vuetify/components/VTextField";
+import VariableBtn from "@/components/Modules/Common/VariableBtn.vue";
 
 export interface DeploymentSpecificationUpdateEnvironmentVariableDialog_Input {
     environmentVariable: {
@@ -26,58 +27,6 @@ const used = ref(false);
 const showDialog = ref(false);
 const name = ref('');
 const value = ref('');
-
-const showVariablesMenu = ref(false);
-const variables = ref<Variable[]>([
-    {
-        name: "Namespace",
-        code: "${namespace}"
-    },
-    {
-        name: "Database Host",
-        code: "${database.host}"
-    },
-    {
-        name: "Database Name",
-        code: "${database.name}"
-    },
-    {
-        name: "Database User",
-        code: "${database.user}"
-    },
-    {
-        name: "Database Password",
-        code: "${database.pass}"
-    },
-    {
-        name: "Email Service Host",
-        code: "${emailService.host}"
-    },
-    {
-        name: "Email Service Port",
-        code: "${emailService.port}"
-    },
-    {
-        name: "Email Service User",
-        code: "${emailService.user}"
-    },
-    {
-        name: "Email Service Pass",
-        code: "${emailService.pass}"
-    },
-    {
-        name: "Email Service Sender",
-        code: "${emailService.sender}"
-    },
-    {
-        name: "Workspace Id",
-        code: "${workspace.id}"
-    },
-    {
-        name: "Workspace Name",
-        code: "${workspace.name}"
-    },
-]);
 
 // <editor-fold desc="Functions">
 
@@ -157,37 +106,9 @@ function onVariableClicked(variable: Variable) {
                                 spellcheck="false"
                             />
 
-                            <v-menu
-                                v-model="showVariablesMenu"
-                                :close-on-content-click="false"
-                                left
-                                min-width="250"
-                                offset-y>
-                                <template v-slot:activator="{ props }">
-                                    <v-btn
-                                        v-bind="props"
-                                        icon
-                                        variant="plain"
-                                        color="primary"
-                                        size="small">
-                                        <v-icon>fa fa-plus</v-icon>
-                                        <v-tooltip activator="parent" location="bottom">Insert variable</v-tooltip>
-                                    </v-btn>
-                                </template>
-
-                                <v-list
-                                    class="list-items">
-                                    <v-list-item
-                                        v-for="(variable, i) in variables" :key="i"
-                                        dense
-                                        @click="onVariableClicked(variable)">
-                                        <v-list-item-title>
-                                            <v-icon size="small" class="my-auto">fa fa-window-maximize fa</v-icon>
-                                            <span class="ml-2">{{ variable.name }}</span>
-                                        </v-list-item-title>
-                                    </v-list-item>
-                                </v-list>
-                            </v-menu>
+                            <variable-btn
+                                @add-variable="item => value += item"
+                            />
                         </div>
                     </v-col>
                 </v-row>